Refactor : refactor codes based on reviewer's review (2)
