+++ src/main/java/ladder/model/row/Position.java
@@ -1,7 +1,6 @@
-package ladder.model.player;
+package ladder.model.row;
 
-import ladder.model.row.Row;
-import ladder.model.row.Rows;
+import ladder.model.player.Direction;
 
 import java.util.Objects;
 
@@ -15,16 +14,12 @@ public class Position {
         this.position = position;
     }
 
-    public static Position of(int position){
+    public static Position of(int position) {
         return new Position(position);
     }
 
     public int getPosition() {
         return position;
-    }
-
-    public Position add(int value){
-        return new Position(position + value);
     }
 
     public Position findFinalLocation(Rows rows) {
