+++ src/main/java/ladder/model/row/Row.java
@@ -1,6 +1,5 @@
 package ladder.model.row;
 
-import ladder.model.player.Position;
 import ladder.randomgenerator.BooleanGenerator;
 
 import java.util.HashMap;
