+++ src/main/java/ladder/view/ResultView.java
@@ -3,7 +3,7 @@ package ladder.view;
 import ladder.model.player.Player;
 import ladder.model.player.PlayerName;
 import ladder.model.player.Players;
-import ladder.model.player.Position;
+import ladder.model.row.Position;
 import ladder.model.prize.LadderPrizes;
 import ladder.model.result.GameResult;
 import ladder.model.row.Row;
@@ -20,35 +20,20 @@ public class ResultView {
     private static final String RESULT_TITLE = "실행 결과";
     private static final int ONE_BLOCK_WIDTH = 6;
     private static final double HALF_BLOCK_WIDTH = 2.5;
+//
+//    public ResultView(GameResult gameResult) {
+//        this.gameResult = gameResult;
+//    }
 
-    private Players players;
-    private Rows rows;
-    private LadderPrizes ladderPrizes;
-    private GameResult gameResult;
+//    public static ResultView of(GameResult gameResult) {
+//        return new ResultView(gameResult);
+//    }
 
-    public ResultView(Players players, Rows rows, LadderPrizes ladderPrizes) {
-        this.players = players;
-        this.rows = rows;
-        this.ladderPrizes = ladderPrizes;
-    }
-
-    public ResultView(GameResult gameResult) {
-        this.gameResult = gameResult;
-    }
-
-    public static ResultView of(GameResult gameResult) {
-        return new ResultView(gameResult);
-    }
-
-    public static ResultView of(Players players, Rows rows, LadderPrizes ladderPrizes) {
-        return new ResultView(players, rows, ladderPrizes);
-    }
-
-    public void printLadder() {
+    public static void printLadder(Players players, Rows rows, LadderPrizes ladderPrizes) {
         printTitle();
-        printPlayers();
-        printRows();
-        printPrizes();
+        printPlayers(players);
+        printRows(rows, players);
+        printPrizes(ladderPrizes);
         printBlankLine();
     }
 
@@ -58,17 +43,17 @@ public class ResultView {
         printGameResult(gameResult);
     }
 
-    public Players getPlayers() {
-        return players;
-    }
-
-    public Rows getRows() {
-        return rows;
-    }
-
-    public LadderPrizes getLadderPrizes() {
-        return ladderPrizes;
-    }
+//    public Players getPlayers() {
+//        return players;
+//    }
+//
+//    public Rows getRows() {
+//        return rows;
+//    }
+//
+//    public LadderPrizes getLadderPrizes() {
+//        return ladderPrizes;
+//    }
 
     private static void printGameResult(GameResult gameResult) {
         gameResult.getKeySet().stream()
@@ -81,7 +66,7 @@ public class ResultView {
         System.out.println(RESULT_TITLE);
     }
 
-    private void printPrizes() {
+    private static void printPrizes(LadderPrizes ladderPrizes) {
         ladderPrizes.getLadderPrizes().stream()
                 .map(it -> it.getPrizeName())
                 .map(it -> it.getPrizeName())
@@ -90,13 +75,13 @@ public class ResultView {
         printBlankLine();
     }
 
-    private void printTitle() {
+    private static void printTitle() {
         System.out.println(BORDER_LINE_TOP);
         System.out.println(LADDER_TITLE);
         printBlankLine();
     }
 
-    private void printPlayers() {
+    private static void printPlayers(Players players) {
         players.getPlayers().stream()
                 .map(Player::getName)
                 .map(PlayerName::getName)
@@ -105,21 +90,21 @@ public class ResultView {
         printBlankLine();
     }
 
-    private void printRows() {
+    private static void printRows(Rows rows, Players players) {
         for (int i = 0; i < rows.getSize(); i++) {
-            printOneRow(rows.getRowsElement(i));
+            printOneRow(rows.getRowsElement(i), players);
         }
     }
 
-    private int findBlankCountBefore(String name) {
+    private static int findBlankCountBefore(String name) {
         return (int) Math.ceil((ONE_BLOCK_WIDTH - name.length()) / HALF_BLOCK_WIDTH);
     }
 
-    private int findBlankCountAfter(String name) {
+    private static int findBlankCountAfter(String name) {
         return (int) Math.round((ONE_BLOCK_WIDTH - name.length()) / HALF_BLOCK_WIDTH);
     }
 
-    private String createBlanks(int count) {
+    private static String createBlanks(int count) {
         StringBuilder stringBuilder = new StringBuilder();
         for (int i = 0; i < count; i++) {
             stringBuilder.append(ONE_BLANK);
@@ -127,7 +112,7 @@ public class ResultView {
         return stringBuilder.toString();
     }
 
-    private void printOneRow(Row row) {
+    private static void printOneRow(Row row, Players players) {
         printBlankPriorToRow();
 
         System.out.print(ONE_VERTICAL);
@@ -138,19 +123,19 @@ public class ResultView {
         printBlankLine();
     }
 
-    private void printBlankPriorToRow() {
+    private static void printBlankPriorToRow() {
         for (int i = 0; i < 2; i++) {
             System.out.print(ONE_BLANK);
         }
     }
 
-    private void printValue(boolean booleanValue) {
+    private static void printValue(boolean booleanValue) {
         String input = booleanValue ? BLOCK_TRUE : BLOCK_FALSE;
         System.out.print(input);
         System.out.print(ONE_VERTICAL);
     }
 
-    private void printBlankLine() {
+    private static void printBlankLine() {
         System.out.println();
     }
