+++ src/test/java/ladder/domain/player/PlayerTest.java
@@ -1,7 +1,7 @@
 package ladder.domain.player;
 
 import ladder.model.player.Player;
-import ladder.model.player.Position;
+import ladder.model.row.Position;
 import ladder.model.row.Row;
 import ladder.model.row.Rows;
 import org.junit.jupiter.api.BeforeEach;
